==> 01_Data_Mining.ipynb <==
# Data Mining from Youtube Channels

## Setup
Import required packages including Google's API for Youtube. For more information please see: https://developers.google.com/youtube/v3

import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
from yt_api_jb import YtScraper

Here we'll look at my brother's Youtube channel, WhyISalty.

A Google Account is required to access the Google API Console. For this project, an application was created using the [Google Developers Console](https://console.developers.google.com/).

Now let's import the YtScraper class and create an object that contains information and methods related to the WhyISalty channel:

# Call the Youtube API Scraper and list all videos WhyISalty's Youtube channel
yt_whyisalty = YtScraper()
video_ids = yt_whyisalty.get_upload_ids("WhyISalty")
print(video_ids[0:6]) # First 5 video ids

Each ID above corresponds to a Video or Live Stream uploaded by the User. We can also get the total statistics across all videos:

channel_metrics_df = yt_whyisalty.metrics
print(channel_metrics_df)

## Data Mining

video_metric_df = yt_whyisalty.get_video_metrics(video_ids)
print(video_metric_df.head())

video_comments_df = yt_whyisalty.get_video_comments(video_ids)
print(video_comments_df.head())

## Database Creation

Now we'll export both of these tables as separate CSV files. We'll also create a single row CSV file containing the channel metrics to better represent relational databases.

video_metric_df.to_csv('csv/video_metrics.csv')
video_comments_df.to_csv('csv/video_comments.csv')
