==> 02_Data_Cleaning.ipynb <==
# Data Cleaning from Gathered Data

## Setup

Here we'll use some of the same libraries to clean the data gathered from the YouTube API.

import pandas as pd
import numpy as np

video_metrics_df = pd.read_csv('csv/video_metrics.csv', index_col=0)
video_comments_df = pd.read_csv('csv/video_comments.csv', index_col=0)
print(f'video_metrics_df shape: {video_metrics_df.shape}')
print(f'video_comments_df shape: {video_comments_df.shape}')

## Cleaning

First let's look at the video_metrics_df

video_metrics_df.dtypes

Based on the above, it looks like that most columns are the expected type; numerical values are integers, strings are objects. 'published_at' and 'duration' should ideally be datetime format, but since we plan to upload these cleaned CSV files to a SQL database we'll refrain for now. 'category_id' should be an object rather than an integer since these values correspond with a separate table that links each category_id to a category name.

video_metrics_df['category_id'] = video_metrics_df['category_id'].astype('str')
print(video_metrics_df['category_id'].describe())

Now if we look at the categorical features further.

video_metrics_df.describe(include=['O'])

We can see that most columns have unique entries, as each row represents a unique video. We also expect the tags, description, and category IDs to be re-used at times. Interestingly the title of the video appears to be duplicated for a few videos - let's explore why that may be the case.

duplicated_titles = video_metrics_df[video_metrics_df['title'].duplicated()]['title']
print(duplicated_titles.to_list())
video_metrics_df[video_metrics_df['title'].isin(duplicated_titles)]

Nothing unusual here, all duplicated titles have different values in other columns. 

The only exception is the first set of videos which share the same title and description, but since these were from steams on back-to-back days most likely the title was re-used.

Finally let's look for missing values.

print(video_metrics_df.isna().sum())

video_metrics_df[video_metrics_df['description'].isna()].head()

Doing a quick check on YouTube shows that NaN values correspond with videos that had no description provided. We'll update this as 'No description provided' to remove any confusion.

video_metrics_df['description'] = video_metrics_df['description'].replace(np.nan, 'No description provided')
print(video_metrics_df['description'].describe())

Now let's repeat the process with the video_comments_df.

video_comments_df.dtypes

video_comments_df.describe()

print(video_metrics_df.isna().sum())

Everything looks good! Lets save these tables and ingest into a SQL database for subsequent analysis

video_comments_df.to_csv('csv/video_comments_cleaned.csv')
video_metrics_df.to_csv('csv/video_metrics_cleaned.csv')
