==> 04_Feature_Engineering.ipynb <==
# Feature Engineering after Data Exploration

After exploring some of the data, there remained several metrics that we were unable to explore due to formatting such as:
- Text-based data with unique entries
- Time-based data
- Unstructed image-based data

Therefore, it may now be worthwhile to generate new features that summarize the above metrics easier.

## Setup

# Setup API
from google.cloud import bigquery
from google.oauth2 import service_account
import pandas as pd
import config as cfg
import os

sa_credentials = 'gpc/youtube-scraper-404402-a6dc21ea107a.json' # download from service_account page
os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = sa_credentials

credentials = service_account.Credentials.from_service_account_file(
    sa_credentials, scopes=["https://www.googleapis.com/auth/cloud-platform"]
)

bq_client = bigquery.Client(credentials=credentials, project=credentials.project_id)

# Read tables into python
query_string = """
SELECT 
    *
FROM youtube-scraper-404402.yt_data.video_metrics_cleaned
"""
video_metrics_df = bq_client.query(query_string).to_dataframe()

query_string = """
SELECT 
    *
FROM youtube-scraper-404402.yt_data.video_comments_cleaned
"""
video_comments_df = bq_client.query(query_string).to_dataframe()

video_metrics_df.head()

video_comments_df.head()

## Text-based Features

For both tables, let's look at how we can extract meaninful information from:
- Video Title
- Video Description
- Video Tags
- Video Comments

### Video Title

Let's look at the first 10 video titles:

for title in video_metrics_df['title'].head(10):
    print(title)

It seems like videos might contain the subject matter in the title (such as which video game is being played), let's explore that in more detail.

video_titles = video_metrics_df['title']
subject_matter = []
for title in video_titles:
    if ' | ' in title:
        subject_matter.append(title.split(' | ')[1])
    else:
        subject_matter.append('None')
video_metrics_df['subject_matter'] = subject_matter
video_metrics_df.value_counts('subject_matter')


We may need to reformat some of these for consistency, but let's also make sure that we're not missing anything for titles labeled as 'None'

video_metrics_df[video_metrics_df['subject_matter'] == 'None']['title']

It looks like that other videos contain the subject matter as hashtags or enclosed in square brackets. Since there are multiple hashtags in a title, we'll need to ingnores those are try to look in the tags section insted, however let's try to pull out those enclosed in square brackets.

import re

video_titles = video_metrics_df['title']
subject_matter = []
for title in video_titles:
    if ' | ' in title:
        subject_matter.append(title.split(' | ')[1])
    elif '[' in title and ']' in title:
        subject_matter.append(re.sub('.*.\[', '', re.sub('\].*.|\]', '', title)))
    else:
        subject_matter.append('None')
        
video_metrics_df['subject_matter'] = subject_matter
video_metrics_df.value_counts('subject_matter')

Now we've significantly reduced the number of videos with no subject matter as it looks like videos based on the Eternal Card Game were enclosed in square brackets. Again, let's check if there is anything else that needs to be extracted from titles with no subject matter.

video_metrics_df[video_metrics_df['subject_matter'] == 'None']['title'].head(35)

video_metrics_df[video_metrics_df['subject_matter'] == 'None']['title'].tail(35)

It looks like some videos also have the subject matter prior to the title, seperated by the '-' character. Let's try to grab those and then take a look at the final list.

video_titles = video_metrics_df['title']
subject_matter = []
for title in video_titles:
    if ' | ' in title:
        subject_matter.append(title.split(' | ')[1])
    elif '[' in title and ']' in title:
        subject_matter.append(re.sub('.*.\[', '', re.sub('\].*.|\]', '', title)))
    elif ' - ' in title:
        subject_matter.append(re.sub(' - .*.', '', title))
    else:
        subject_matter.append('None')
        
video_metrics_df['subject_matter'] = subject_matter
video_metrics_df.value_counts('subject_matter')

Let's look at titles with 'None' for subject title one more time and see if we can simply pull keywords.

video_metrics_df[video_metrics_df['subject_matter'] == 'None']['title'].to_list()

Based on the above and the subject matter list, we can pull the following:
- Armored Core
- Mechabellum
- Eternal
- Project Champion
- Mythgard
- Titan Core
- Monster Hunter

There are some typos for some of these titles but we'll leave them alone as that may potentially influence the YouTube algorithm.

video_titles = video_metrics_df['title']
subject_matter = []
for title in video_titles:
    if ' | ' in title:
        subject_matter.append(title.split(' | ')[1])
    elif '[' in title and ']' in title:
        subject_matter.append(re.sub('.*.\[', '', re.sub('\].*.|\]', '', title)))
    elif ' - ' in title:
        subject_matter.append(re.sub(' - .*.', '', title))
    elif 'armored' in title.lower():
        subject_matter.append('Armored Core')
    elif 'mechabellum' in title.lower():
        subject_matter.append('Mechabellum')
    elif 'eternal' in title.lower():
        subject_matter.append('Eternal Card Game')
    elif 'project champion' in title.lower():
        subject_matter.append('Project Champion')
    elif 'mythgard' in title.lower():
        subject_matter.append('Mythgard')
    elif 'titan core' in title.lower():
        subject_matter.append('Titan Core')
    elif 'monster hunter' in title.lower():
        subject_matter.append('Monster Hunter')
    else:
        subject_matter.append('None')
        
video_metrics_df['subject_matter'] = subject_matter
video_metrics_df.value_counts('subject_matter')

One last sanity check for videos with no subject matter in the title:

video_metrics_df[video_metrics_df['subject_matter'] == 'None']['title'].to_list()

Great! Let's also clean up the subject matter for topics with 5 or more videos, the rest we can classify as 'Misc.'.

# Clean up variations of each subject matter
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Eternal'), 'subject_matter'] = 'Eternal Card Game'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Armored Core'), 'subject_matter']= 'Armored Core'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Titan Core'), 'subject_matter'] = 'Titan Core'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Mechabellum'), 'subject_matter'] = 'Mechabellum'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Project Champion'), 'subject_matter']= 'Project Champions'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Monster Hunter'), 'subject_matter'] = 'Monster Hunter'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.lower().str.contains('mtg arena'), 'subject_matter'] = 'MTG Arena'
video_metrics_df.loc[video_metrics_df['subject_matter'].str.contains('Hearthstone'), 'subject_matter'] = 'Hearthstone'

# Get subject matter in 5 or more videos
subject_matter_count = video_metrics_df['subject_matter'].value_counts()
popular_subjects = subject_matter_count[subject_matter_count >= 5].index.to_list()

video_metrics_df.loc[~video_metrics_df['subject_matter'].isin(popular_subjects), 'subject_matter'] = 'Misc.'

video_metrics_df['subject_matter'].value_counts()

video_metrics_df.loc[video_metrics_df['subject_matter'] == 'STREAM', 'title']

The STREAM subject matter seems to be a side effect of the square bracket format, let's assign those the appropriate subject matter manually:

video_metrics_df.loc[video_metrics_df['subject_matter'] == 'STREAM', 'subject_matter'] = ['None', 'Eternal Card Game', 'Titan Core', 'Monster Hunter', 'Titan Core', 'Monster Hunter', 'Armored Core']
video_metrics_df['subject_matter'].value_counts()

While we're here, let's make a separate feature based on whether the video was streamed or not based on the title.

video_metrics_df['streamed'] = [1 if 'STREAM' in x.upper() else 0 for x in video_metrics_df['title']]
video_metrics_df['streamed'].value_counts()

### Video Description

Let's take a look at the first 10 video description entries:

video_metrics_df['description'].head(10).to_list()

Based on this, there's a couple of things we can do:
- Remove hashtags and special characters to clean the text for downstream NLP sentiment analysis.
- Create three features based on the presence of hashtags, timestamps, and socials

video_metrics_df['description_text_only'] = [re.sub('#.* |#.*|\\n|\\\'|([0-9]|[0-9][0-9]):.*.', '', x) for x in video_metrics_df['description']]
video_metrics_df['description_hashtags'] = [1 if '#' in x else 0 for x in video_metrics_df['description']]
video_metrics_df['description_timestamps'] = [1 if re.search('([0-9]|[0-9][0-9]):', x) else 0 for x in video_metrics_df['description']]
video_metrics_df['description_socials'] = [1 if re.search('discord|patreon|twitch|twitter', x.lower()) else 0 for x in video_metrics_df['description']]

Now let's confirm each one:

video_metrics_df['description_text_only'].head().to_list()

video_metrics_df[video_metrics_df['description_hashtags'] == 1]['description'].head()

video_metrics_df[video_metrics_df['description_timestamps'] == 1]['description'].head().to_list()

video_metrics_df[video_metrics_df['description_socials'] == 1]['description'].head().to_list()

### Video Comments

Similar to the Title and Description, let's also look at the comments.

video_comments_df['video_comment'].head(20).to_list()

video_comments_df['video_comment'].tail().to_list()

Similar to the video description, we need to remove the special characters. It also seems that during the data cleaning step, that we accidently introduced some NoneType variables - let's remove those rows since they represent videos without comments.

video_comments_df = video_comments_df.dropna()
video_comments_df.loc[:,'video_comment'] = [re.sub('\\n|\\\'', '', x) for x in video_comments_df['video_comment']]

video_comments_df['video_comment'].head(20).to_list()

video_comments_df['video_comment'].tail().to_list()

## Time-based Data

For both tables, there are two metrics that can be further extracted as independent features:
- Year, Month, Day of the Week of Upload
- Hour of Upload
- Video Duration in Seconds
- Video Comment Time Relative to Upload in Seconds

Let's start with the first three features in the video_metrics_df. 

For Year, Month, Day of the Week Upload, Hour of Upload we'll use the published_at column which is in YYYY-MM-DD HH-MM-SS TimeZone format. Luckily this is already in 'datetime64' by Pandas. We an use the datetime library to extract the information we need.

import datetime
#[x.]
print('Year : ', video_metrics_df['published_at'][0].year)
print('Month : ', video_metrics_df['published_at'][0].month)
print('Weekday : ', video_metrics_df['published_at'][0].weekday())
print('Hour : ', video_metrics_df['published_at'][0].hour)

video_metrics_df['published_year'] = [x.year for x in video_metrics_df['published_at']]
video_metrics_df['published_month'] = [x.month for x in video_metrics_df['published_at']]
video_metrics_df['published_weekday'] = [x.weekday() for x in video_metrics_df['published_at']]
video_metrics_df['published_hour'] = [x.hour for x in video_metrics_df['published_at']]

The duration of the video can be found under the 'duration' column and follows a slightly trickier format 'PTn(n)Hn(n)Mn(n)S where not all numbers and characters are present. If the duration is 22 seconds long, the format is 'PT22S'. If the duration is 8 minutes and 37 seconds long, the format is 'PT8M37S'.

Let try to write a function that will start from the right of the 'PT' pattern, pull the numeric values to the right of each character, and use the character pattern (H/M/S) to convert the following numeric value to seconds.

# example of a successful regex pattern
test = 'PT8M37S'
print(re.findall(r'[0-9]{1,2}[A-Z]', test))

def pt_to_seconds(pt_string):
    '''Converts string in PT(n(n)H)(n(n)M)n(n)S) format to seconds, where () enclose optional values and H/M/S denote hours/minutes/seconds'''
    # use regex to split each string by their right most character
    pt_seconds = 0
    time_split = re.findall(r'[0-9]{1,2}[A-Z]', pt_string)
    # convert each time format to seconds and append
    for split in time_split:
        num_split = int(split[0:-1])
        if 'H' in split:
            pt_seconds += num_split * 360
        elif 'M' in split:
            pt_seconds += num_split * 60
        else:
            pt_seconds += num_split
    return(pt_seconds)

Let's test the string 'PT8M37S' which should return 8*60 + 37 = 517

pts_example = pt_to_seconds('PT8M37S')
print(pts_example, ' seconds')

video_metrics_df['video_duration_seconds'] = video_metrics_df['duration'].apply(lambda x: pt_to_seconds(x))
print(video_metrics_df['video_duration_seconds'])

Finally let's add the feature 'comment_relative_to_upload_seconds'. This will be a simple subtraction of the video_comment_time column in the video_comments_df with the published_at column in the video_metrics_df.

Because both columns have entries already in datetime format, we can use the datetime library to perform this operation.

Similar to above, let's write a function that uses the video_id from each comment to pull the published_at value from the video_metrics_df.

def comment_from_upload_seconds(row):
    '''extracts the video_id and datetime values from the row and finds thier respective values in the comment_metrics_df dataframe. Subtracts from both datetime values to get the time from upload'''
    c_video_id = row['video_id']
    c_datetime = row['video_comment_time']
    v_datetime = video_metrics_df[video_metrics_df['video_id'] == c_video_id]['published_at']
    c_from_v = c_datetime - v_datetime
    c_from_v_seconds = c_from_v.iloc[0].seconds
    return c_from_v_seconds

Let's validate this using the comment_id 'UgzWhoTNYW-C1nD1Lm94AaABAg' and 'Ugx5R_BBSeGsBuC3AN54AaABAg'.

row_1 = video_comments_df[video_comments_df['video_comment_parent_id'] == 'UgzWhoTNYW-C1nD1Lm94AaABAg']
row_2 = video_comments_df[video_comments_df['video_comment_parent_id'] == 'Ugx5R_BBSeGsBuC3AN54AaABAg']

print('Comment 1:', row_1['video_comment_time'].iloc[0])
print('Comment 2:', row_2['video_comment_time'].iloc[0])

row_1_video = video_metrics_df[video_metrics_df['video_id'] == row_1['video_id'].iloc[0]]
row_2_video = video_metrics_df[video_metrics_df['video_id'] == row_2['video_id'].iloc[0]]

print('Comment 1 video_id:', row_1_video['video_id'].iloc[0], 'and datetime', row_1_video['published_at'].iloc[0])
print('Comment 2 video_id:', row_2_video['video_id'].iloc[0], 'and datetime', row_2_video['published_at'].iloc[0])

row_1_sub = (row_1['video_comment_time'].iloc[0] - row_1_video['published_at'].iloc[0]).seconds
row_2_sub = (row_2['video_comment_time'].iloc[0] - row_2_video['published_at'].iloc[0]).seconds

print('Comment 1 expected output', row_1_sub, 'seconds')
print('Comment 2 expected output', row_2_sub, 'seconds')

video_comments_df[video_comments_df['video_comment_parent_id'].isin(['UgzWhoTNYW-C1nD1Lm94AaABAg', 'Ugx5R_BBSeGsBuC3AN54AaABAg'])].apply(lambda x:comment_from_upload_seconds(x), axis = 1)

We get the expected output, so let's apply that to all rows and append it to the video_comments_df:

video_comments_df['comments_relative_to_upload_seconds'] = video_comments_df.apply(lambda x: comment_from_upload_seconds(x), axis = 1)

## Sentiment Analysis

Let's finally perform some simple sentiment analysis on the video titles as well as comments.

Here we'll use the Natural Language Toolkit (NLTK) library which requires the following pre-processing steps prior to modelling:
1. Tokenization
2. Lowercasing
4. Noise removal
5. Spelling correction
6. Lemmatization
7. Part-of-speech tagging
8. Named entities recognition
9. Text normalization

import nltk
nltk_comments = video_comments_df[['video_comment']]

Before we write a function which will preprocess the text, let's validate each step.

### Tokenization and Lowercasing

First let's split the text into individual words, also known as tokens:

tokens_list = [nltk.word_tokenize(x.lower()) for x in nltk_comments['video_comment']]
print(tokens_list[0])

### Noise removal

import string

# Remove punctuation, stop words, whitespace, urls, html code
stop_words = nltk.corpus.stopwords.words('english')
url_pattern = r"(http|ftp|https)://([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-])?"
html_pattern = r"<[^>]+>"

filtered_token_list = []
for tokens in tokens_list:
    filtered_tokens = [x for x in tokens if x not in string.punctuation and x not in stop_words]
    filtered_tokens = [" ".join(x.strip().split()) for x in filtered_tokens]
    filtered_tokens = [re.sub(html_pattern, '', re.sub(url_pattern, '', x)) for x in filtered_tokens]
    filtered_token_list.append(filtered_tokens)

print(filtered_token_list[0])
